Transform Citizens Into Demons tests pass.
- fix turn expectation for Mr. Sinister; there is not even 1 Bystander
  for him to capture with this scheme

- count the 'other' cards among total villain cards

Jean Grey cards with this scheme are still not *strictly* counted as
villain cards. That is, if there are so many Villain cards that play the
top card of the Villain Deck if it is a Villain (i.e., Egghead), they
will not count Jean Grey and the minimum turn that the Villain Deck
could run out will be incorrect. This may not ever matter, so don't
implement for now.